Improve root import tree shaking
